+++ src/security/groups/group-operations.spec.ts
@@ -451,7 +451,7 @@ describe('Group Operations', () => {
 
             expect(users).not.toBeNull();
             expect(users.items.length).toEqual(2);
-            expect(users.items[1].userName).toEqual(defaultUsers[1].userName);
+            expect(users.items[1].username).toEqual(defaultUsers[1].username);
         });
 
         it('by group name', async () => {
@@ -467,7 +467,7 @@ describe('Group Operations', () => {
 
             expect(users).not.toBeNull();
             expect(users.items.length).toEqual(2);
-            expect(users.items[1].userName).toEqual(defaultUsers[1].userName);
+            expect(users.items[1].username).toEqual(defaultUsers[1].username);
         });
     });