+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-beta.5] - 2021-04-01
+### Changed
+- users: renamed user properties to 'username', 'firstname' and 'lastname'
+- roles: change role 'name' and 'description' properties to string
+
+## [1.0.2-beta.4] - 2021-03-12
+### Changed
+- updated reference to contensis-core-api
+
 ## [1.0.2-beta.4] - 2021-03-12
 ### Changed
 - updated reference to contensis-core-api
@@ -42,7 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0-rc.3] - 2020-08-13
 
-### Chnaged
+### Changed
 
 - groups: updated path for adding and removing child groups;